Expanding the range of Pell numbers

+++ src/ru/mazila/ylab/Pell.java
@@ -13,7 +13,7 @@ public class Pell {
         }
     }
 
-    public static int redo(int n) {
+    public static long redo(int n) {
 
         if (n == 0) return 0;
         if (n == 1) return 1;
